[FIX] Duplicate variant check Fix.
1) If variant stored has more attributes (e.g. not all required)
and a new variant was being proposed, it was matching.  Need to
reject variants with more attributes as not being a match!

2) Likewise, custom value ids MUST be taken in to account.  If
a variant had Attr1: Value 1, Attr2: Value X, and a new variant
was being checked with Attr1: Value 1, Attr2: Custom, then it was
matching and not creating a new variant!