[FIX] Readonly fields not being stored.
If the wizard turns a field to readony because the combination
means it is unenterable now, then the value needs to be submitted
when the client attempts to write the record.

Since the Odoo client does not submit readonly values, then we
need to ensure we get that value by keeping an "invisible" equivalent
of the real value, which is never readonly.