[FIX] MRP BOM creation
The configurator was creating a new BOM, every time
a configuration was retrieved!  Routine now only
creates a BOM if there is none.

Also, refactored to allow for overriding and hooks.
This allows custom code to create a default BOM before
the attributes are added, and for more clever adding
of products (quantities, etc).